user login & logout

+++ base/views.py
@@ -1,6 +1,10 @@
 from django.shortcuts import render, redirect
+from django.http import HttpResponse
+from django.contrib import messages
+from django.contrib.auth.decorators import login_required
+from django.contrib.auth import authenticate, login, logout
 from django.db.models import Q
-from .models import Room, Topic
+from .models import Room, Topic, User
 from .forms import RoomForm
 # Create your views here.
 
@@ -10,6 +14,34 @@ from .forms import RoomForm
 #     {'id':3, "name":"Real Estate"},
 #     {'id':4, "name":"Education"}
 # ]
+
+
+def loginPage(request):
+
+    if request.user.is_authenticated:
+        return redirect('home')
+
+    username = request.POST.get('username')
+    password = request.POST.get('password')
+    # try:
+    #     user = User.objects.get(username=username)
+    # except:
+    #     messages.error(request, 'User does not exists.')
+
+    user = authenticate(request, username=username, password=password)
+
+    if user is not None:
+        login(request, user)
+        return redirect('home')
+    else:
+        messages.error(request, 'Username OR password does not exist')
+     
+    context = {}
+    return render(request, 'base/login_register.html', context)
+
+def logoutUser(request):
+    logout(request)
+    return redirect('home')
 
 def home(request):
     q = request.GET.get('q') if request.GET.get('q') != None else ''
@@ -30,6 +62,7 @@ def room(request, pk):
     context = {'room':room}
     return render(request, 'base/room.html', context)
 
+@login_required(login_url='/login')
 def createRoom(request):
     form = RoomForm()
     if request.method == 'POST':
@@ -41,9 +74,13 @@ def createRoom(request):
     return render(request, 'base/room_form.html', context)
 
 
+@login_required(login_url='/login')
 def updateRoom(request, pk):
     room = Room.objects.get(id = pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse('You are not allowed here!!')
     
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
@@ -55,10 +92,13 @@ def updateRoom(request, pk):
     return render(request, 'base/room_form.html',context)
 
 
-
+@login_required(login_url='/login')
 def deleteRoom(request, pk):
     room =Room.objects.get(id = pk)
     form = RoomForm(instance = room)
+
+    if request.user != room.host:
+        return HttpResponse('You are not allowed here!!')
 
     if request.method == 'POST':
         room.delete()
